pyproject.toml
[tool.setuptools]
include-package-data = true


[project]
name = "rds-tunnel"
version = "1.0.0"
description = "Simple SSH Tunnel tool for connecting to RDS. Supports MySQL and PostgreSQL."
authors = [
    { name = "Sebastian Bekx", email = "sbekx@cameramatics.com" },
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.40.8",
    "mysql-connector-python>=9.4.0",
    "paramiko==2.11.0",
    "sshtunnel>=0.4.0",
]
[project.scripts]
rds-tunnel = "rds_tunnel.tunnel:cli"
rdst = "rds_tunnel.cli:cli"
